refactor: HypothesisBuilder component
